requirements.txt
asgiref==3.2.10
Django==3.0.8
django-email-utils==0.1.1
django-email-verification==0.0.5.post1
django-filter==2.3.0
django-rest-email-auth==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
hdlConvertorAst==0.2
hwt==3.2
ipCorePackager==0.5
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.2.0
pycocotb==0.6
pyDigitalWaveTools==0.5
PyJWT==1.7.1
pyMathBitPrecise==0.5
pytz==2020.1
sortedcontainers==2.2.2
sqlparse==0.3.1
